pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = ""
authors = [
    {name = "Seyeon Kim",email = "ksyrobot@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "langchain-upstage (>=0.6.0,<0.7.0)",
    "langchain-core (>=0.3.30,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "numpy (==1.26.4)",
    "tiktoken (>=0.8.0,<0.9.0)",
    "langchain-text-splitters (>=0.3.5,<0.4.0)",
    "langchain-community (>=0.3.15,<0.4.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "onnxruntime (<1.19.2)",
    "chromadb (>=0.6.3,<0.7.0)",
    "faiss-gpu (>=1.7.2,<2.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pandas (==2.2.3)",
    "apache-airflow (==2.9.3)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "torch (>=2.6.0,<3.0.0)",
    "transformers (<4.48)",
    "pydub (>=0.25.1,<0.26.0)",
    "ffmpeg (>=1.4,<2.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "locust (>=2.32.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"